+++ package.json
@@ -4,7 +4,7 @@
   "description": "YP Sprint 11 project",
   "main": "script.js",
   "scripts": {
-    "build": "rimraf dist && webpack --mode production", // добавили удаление папки
+    "build": "rimraf dist && webpack --mode production",
     "dev": "webpack-dev-server --mode development --open --watch"
   },
   "repository": {
@@ -25,7 +25,9 @@
     "babel-loader": "^8.0.6",
     "css-loader": "^3.2.0",
     "cssnano": "^4.1.10",
+    "file-loader": "^4.2.0",
     "html-webpack-plugin": "^3.2.0",
+    "image-webpack-loader": "^5.0.0",
     "mini-css-extract-plugin": "^0.8.0",
     "path": "^0.12.7",
     "postcss-loader": "^3.0.0",
