+++ readme.md
@@ -1,8 +1,10 @@
 Яндекс.Практикум Когорта 1. Александр Демин. Проект 11 спринта v1.0.1
 =====================================================================
 
-
+https://rationabilis.github.io/Jaques/
 ------------------------------------------------------
+
+Исходный JS код разбит на модули.
