+++ src/blocks/validateInput/validateInput.js
@@ -1,0 +1,7 @@
+/* Функции валидации ввода*/
+
+import validateForm from '../validateForm/validateForm';
+
+export default function validateInput(evt) {
+    validateForm(evt.target.parentElement);
+  }
