+++ src/style.css
@@ -1,0 +1,1 @@
+@import url(./pages/index.css);
