+++ webpack.config.js
@@ -1,7 +1,7 @@
 const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const WebpackMd5Hash = require('webpack-md5-hash');
-const MiniCssExtractPlugin = require("mini-css-extract-plugin"); // добавили плагин
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 module.exports = {
     entry: { main: './src/index.js' },
     output: {
@@ -19,7 +19,21 @@ module.exports = {
             },
             {
                 test: /\.css$/,
-                use:  [MiniCssExtractPlugin.loader, 'css-loader', 'postcss-loader'] // добавили минификацию CSS
+                use:  [MiniCssExtractPlugin.loader, 'css-loader', 'postcss-loader']
+            },
+            {
+                test: /\.(png|jpg|gif|ico|svg)$/,
+                use: [
+                        'file-loader?name=../images/[name].[ext]',
+                        {
+                                loader: 'image-webpack-loader',
+                                /* options: { // ... } */
+                        },
+                ],
+            },
+            {
+                test: /\.(eot|ttf|woff|woff2)$/,
+                loader: 'file-loader?name=./vendor/[name].[ext]'
             }
         ]
     },
